requirements-test.txt
pytest==7.4.3
pytest-django==4.7.0
pytest-mock==3.12.0
factory-boy==3.3.0
coverage==7.3.2
pytest-cov==4.1.0